+++ src/Module/BaseSearchModule.php
@@ -5,83 +5,38 @@ namespace Friendica\Module;
 use Friendica\BaseModule;
 use Friendica\Content\ContactSelector;
 use Friendica\Content\Pager;
-use Friendica\Content\Widget;
 use Friendica\Core\L10n;
 use Friendica\Core\Renderer;
+use Friendica\Object\Search\ResultList;
 use Friendica\Util\Proxy as ProxyUtils;
-use Friendica\Util\Strings;
 use Friendica\Model;
 
 /**
- * Multi search module, which is needed for further search operations
+ * Base class for search modules
  */
-class DirectorySearch extends BaseModule
+abstract class BaseSearchModule extends BaseModule
 {
-	public static function content()
+	/**
+	 * Prints a human readable search result
+	 *
+	 * @param ResultList $results
+	 * @param Pager      $pager
+	 * @param string     $header
+	 *
+	 * @return string The result
+	 * @throws \Friendica\Network\HTTPException\InternalServerErrorException
+	 * @throws \ImagickException
+	 */
+	protected static function printResult(ResultList $results, Pager $pager, $header = '')
 	{
-		if (!local_user()) {
-			notice(L10n::t('Permission denied.'));
-			return Login::form();
-		}
-
-		$a = self::getApp();
-
-		if (empty($a->page['aside'])) {
-			$a->page['aside'] = '';
-		}
-
-		$a->page['aside'] .= Widget::findPeople();
-		$a->page['aside'] .= Widget::follow();
-
-		return self::performSearch();
-	}
-
-	public static function performSearch($prefix = '')
-	{
-		$a      = self::getApp();
-		$config = $a->getConfig();
-
-		$community = false;
-
-		$localSearch = $config->get('system', 'poco_local_search');
-
-		$search = $prefix . Strings::escapeTags(trim(defaults($_REQUEST, 'search', '')));
-
-		if (!$search) {
-			return '';
-		}
-
-		$header = '';
-
-		if (strpos($search, '@') === 0) {
-			$search  = substr($search, 1);
-			$header  = L10n::t('People Search - %s', $search);
-			$results = Model\Search::searchUser($search);
-		}
-
-		if (strpos($search, '!') === 0) {
-			$search    = substr($search, 1);
-			$community = true;
-			$header    = L10n::t('Forum Search - %s', $search);
-		}
-
-		$pager = new Pager($a->query_string);
-
-		if ($localSearch && empty($results)) {
-			$pager->setItemsPerPage(80);
-			$results = Model\Search::searchLocal($search, $pager->getStart(), $pager->getItemsPerPage(), $community);
-
-		} elseif (strlen($config->get('system', 'directory')) && empty($results)) {
-			$results = Model\Search::searchDirectory($search, $pager->getPage());
-			$pager->setItemsPerPage($results->getItemsPage());
-		}
-
 		if (empty($results) || empty($results->getResults())) {
 			info(L10n::t('No matches') . EOL);
 			return '';
 		}
-		$id = 0;
 
+		$a = self::getApp();
+
+		$id = 0;
 		$entries = [];
 		foreach ($results->getResults() as $result) {
